+++ src/Cache/CacheDirective.php
@@ -71,7 +71,6 @@ GRAPHQL;
                     ? data_get($root, $rootCacheKey)
                     : null;
                 $fieldName = $resolveInfo->fieldName;
-                $path = $resolveInfo->path;
 
                 /** @var \Illuminate\Cache\TaggedCache|\Illuminate\Contracts\Cache\Repository $cache */
                 $cache = $shouldUseTags
@@ -84,11 +83,9 @@ GRAPHQL;
                 $cacheKey = $this->cacheKeyAndTags->key(
                     $context->user(),
                     $isPrivate,
-                    $parentName,
                     $rootID,
-                    $fieldName,
                     $args,
-                    $path
+                    $resolveInfo
                 );
 
                 // We found a matching value in the cache, so we can just return early